admin

+++ backend/app/Http/Controllers/AdminController.php
@@ -25,7 +25,8 @@ class AdminController extends Controller
         $adminUserCount = User::where('type', 'admin')->count();
         $contentCount = ContentModel::count();
        
-        return view('admin.dashboard', compact('userCount', 'premiumUserCount', 'productionUserCount', 'adminUserCount', 'contentCount'));
+        //return view('admin.dashboard', compact('userCount', 'premiumUserCount', 'productionUserCount', 'adminUserCount', 'contentCount'));
+        return response()->json($userCount, $premiumUserCount, $productionUserCount, $adminUserCount, $contentCount);
         // return view('admin.dashboard', compact('userCount'));
 
         // $users = User::All()->count();
@@ -36,34 +37,39 @@ class AdminController extends Controller
     function viewAllUsers()
     {
         $users = User::all();
-        return view('admin.users.allUser', ['users' => $users]);
-        
+        //return view('admin.users.allUser', ['users' => $users]);
+        return response()->json($users);
     }
 
     function viewAllPremiumUsers()
     {
         $users=User::where('type', 'premium')->get();
-        return view('admin.users.PremiumUser', ['users' => $users]);
+        //return view('admin.users.PremiumUser', ['users' => $users]);
+        return response()->json($users);
     }
     function viewAllProductionHouse()
     {
         $users=User::where('type', 'production')->get();
-        return view('admin.users.ProductionHouse', ['users' => $users]);
+       // return view('admin.users.ProductionHouse', ['users' => $users]);
+        return response()->json($users);
     }
     function viewAlladmin()
     {
         $users=User::where('type', 'admin')->get();
-        return view('admin.users.admin', ['users' => $users]);
+       // return view('admin.users.admin', ['users' => $users]);
+        return response()->json($users);
     }
     function viewAllMovies()
     {
         $ContentModels = ContentModel::all();
-        return view('admin.content')->with('ContentModel',$ContentModels);
+       // return view('admin.content')->with('ContentModel',$ContentModels);
+        return response()->json($ContentModels);
     }
     function viewReport()
     {
         $ReportModel = ReportModel::all();
-        return view('admin.report', ['ReportModel' => $ReportModel]);
+        //return view('admin.report', ['ReportModel' => $ReportModel]);
+        return response()->json($ReportModel);
     }
     function Statistics()
     {
@@ -76,7 +82,8 @@ class AdminController extends Controller
         $ratingCount = RatingModel::count();
         $reportCount = ReportModel::count();
         
-        return view('admin.statistics', compact('userCount', 'premiumUserCount', 'productionUserCount', 'adminUserCount', 'contentCount', 'orderCount', 'ratingCount', 'reportCount'));
+        //return view('admin.statistics', compact('userCount', 'premiumUserCount', 'productionUserCount', 'adminUserCount', 'contentCount', 'orderCount', 'ratingCount', 'reportCount'));
+        return response()->json($userCount, $premiumUserCount, $productionUserCount, $adminUserCount, $contentCount, $orderCount, $ratingCount, $reportCount);
     }
 
 
@@ -84,17 +91,29 @@ class AdminController extends Controller
     function viewUserDetails($id)
     {
         $users = User::where('user_id', $id)->first();
-        return view('admin.users.Details', ['users' => $users]);
+        //return view('admin.users.Details', ['users' => $users]);
+        if ($users) {
+            return response()->json($users);
+        } else {
+            return response()->json(['error' => 'User not found'], 404);
+        }
+
     }
     function profile()
     {
         $user = User::where('email',session()->get('logged'))->first();
-        return view('admin.profile')->with('user',$user);
+        //return view('admin.profile')->with('user',$user);
+        if ($user) {
+            return response()->json($user);
+        } else {
+            return response()->json(['error' => 'User not found'], 404);
+        }
     }
     function payment()
     {
         $OrderModels = OrderModel::all();
-        return view('admin.Users.paymentHistory')->with('OrderModel',$OrderModels);
+        //return view('admin.Users.paymentHistory')->with('OrderModel',$OrderModels);
+        return response()->json($OrderModels);
     }
 
 
@@ -102,7 +121,12 @@ class AdminController extends Controller
     //_____________edit______________
     function editUser($id)
     { $user = User::where('user_id',$id)->first();
-        return view('admin.Users.editUser')->with('user',$user);
+       //return view('admin.Users.editUser')->with('user',$user);
+        if ($user) {
+            return response()->json($user);
+        } else {
+            return response()->json(['error' => 'User not found'], 404);
+        }
     }
    
     //_____________update______________
@@ -131,7 +155,8 @@ class AdminController extends Controller
         $user->type = $request->type;
         $user->status = $request->status;
         $user->save();
-        return redirect()->route('admin.users.all');
+        //return redirect()->route('admin.users.all');
+        return response()->json($user);
     }
     
    
@@ -142,26 +167,38 @@ class AdminController extends Controller
         // $users = User::where('email',session()->get('logged'))->first();
         // return view('admin.users.allUser')->with('users',$users);
         // $user = User::find($id);
-        if($users->email == session()->has('logged'))
-        {
-            session()->flash('msg','User is logged in'); 
-        }
-        else
-        {
-            $users->delete();
-            session()->flash('msg1','User Deleted'); 
-        }
-        $users->delete();
+        // if($users->email == session()->has('logged'))
+        // {
+        //     session()->flash('msg','User is logged in'); 
+        // }
+        // else
+        // {
+        //     $users->delete();
+        //     session()->flash('msg1','User Deleted'); 
+        // }
+        // $users->delete();
 
-        return redirect()->back();
+        // return redirect()->back();
+        if ($users) {
+            $users->delete();
+            return response()->json(['message' => 'User Deleted'], 200);
+        } else {
+            return response()->json(['error' => 'User not found'], 404);
+        }
        
     }
 
     function deleteContent($id)
     {
         $ContentModel = ContentModel::where('content_id', $id)->first();
-        $ContentModel->delete();
-        return redirect()->back();
+        //$ContentModel->delete();
+        //return redirect()->back();
+        if ($ContentModel) {
+            $ContentModel->delete();
+            return response()->json(['message' => 'Content Deleted'], 200);
+        } else {
+            return response()->json(['error' => 'Content not found'], 404);
+        }
     }
 
 
@@ -169,7 +206,8 @@ class AdminController extends Controller
     //_____________create______________
     function createUser()
     {
-        return view('admin.users.createUser');
+        //return view('admin.users.createUser');
+        return response()->json(['message' => 'User Created'], 200);
     } 
     //_____________createSubmit______________
     function createUserSubmit( Request $req)
@@ -202,8 +240,9 @@ class AdminController extends Controller
             $user->type = $req->type;
             $user->email_verified_at = now();
             $user->save();
-            return redirect()->route('admin.users.all');
-    }
+            //return redirect()->route('admin.users.all');
+            return response()->json(['message' => 'User Created'], 200);
+        }
